refactor(quota): move quota model and normalize int16
- move quota item and period types into model (Quota, PeriodType)
- switch quota limits/remaining/usage fields to int16 end-to-end
- update quota mapper and user repositories to use the new model types
- align numeric types with database compatibility
- prepare quota model for reuse in other components

+++ miruzo/internal/adapter/persistence/postgre/user/repository.go
@@ -28,6 +28,6 @@ func (repo repository) GetSingletonUser(
 
 	return persist.User{
 		ID:            user.ID,
-		DailyLoveUsed: uint16(user.DailyLoveUsed),
+		DailyLoveUsed: user.DailyLoveUsed,
 	}, nil
 }
